+++ console2/src/components/molecules/RepositoryList/index.tsx
@@ -20,11 +20,11 @@
 
 import * as React from 'react';
 import { Link } from 'react-router-dom';
-import { Icon, Table } from 'semantic-ui-react';
+import { Table } from 'semantic-ui-react';
 
 import { ConcordKey } from '../../../api/common';
 import { RepositoryEntry } from '../../../api/org/project/repository';
-import { RepositoryActionDropdown } from '../../molecules';
+import { GitHubLink, RepositoryActionDropdown } from '../../molecules';
 
 interface Props {
     orgName: ConcordKey;
@@ -39,41 +39,6 @@ const getSource = (r: RepositoryEntry) => {
     return r.branch;
 };
 
-const gitUrlParse = (s: string): string | undefined => {
-    if (s.startsWith('git')) {
-        // git@gecgithub01.walmart.com:devtools/concord.git
-        const regex = /git@(.*):(.*)\.git/;
-        const match = regex.exec(s);
-        if (!match || match.length !== 3) {
-            return;
-        }
-        return `https://${match[1]}/${match[2]}`;
-    } else if (s.startsWith('http')) {
-        // https://gecgithub01.walmart.com/devtools/concord.git
-        const regex = /http[s]?:\/\/(.*).git/;
-        const match = regex.exec(s);
-        if (!match || match.length !== 2) {
-            return;
-        }
-        return `https://${match[1]}`;
-    }
-
-    return;
-};
-
-const renderGitLink = (s: string) => {
-    const url = gitUrlParse(s);
-    if (!url) {
-        return s;
-    }
-
-    return (
-        <a href={url} target="_blank">
-            {s} <Icon name="external" />
-        </a>
-    );
-};
-
 const renderTableRow = (orgName: ConcordKey, projectName: ConcordKey, row: RepositoryEntry) => {
     return (
         <Table.Row key={row.id}>
@@ -82,7 +47,9 @@ const renderTableRow = (orgName: ConcordKey, projectName: ConcordKey, row: Repos
                     {row.name}
                 </Link>
             </Table.Cell>
-            <Table.Cell>{renderGitLink(row.url)}</Table.Cell>
+            <Table.Cell>
+                <GitHubLink url={row.url} />
+            </Table.Cell>
             <Table.Cell>{getSource(row)}</Table.Cell>
             <Table.Cell>{row.path}</Table.Cell>
             <Table.Cell>{row.secretName}</Table.Cell>